Verify that all templates need to learn from the entire utterance.
